refactor(cockpit): move file stat tests to owner module
