+++ crates/tokmd-cockpit/src/lib.rs
@@ -213,17 +213,4 @@ fn branchy(x: i32) -> i32 {
         assert_eq!(analysis.total_complexity, 4);
         assert_eq!(analysis.max_complexity, 4);
     }
-
-    // ---- FileStat AsRef ----
-
-    #[test]
-    fn test_filestat_as_ref() {
-        let stat = FileStat {
-            path: "src/main.rs".to_string(),
-            insertions: 10,
-            deletions: 5,
-        };
-        let s: &str = stat.as_ref();
-        assert_eq!(s, "src/main.rs");
-    }
 }
